New cli folder structure, install script, circle config
