+++ .gitignore
@@ -40,3 +40,4 @@ Thumbs.db
 **/dist/**/*
 **/public
 gopath
+client/builds/**/*
