+++ Dockerfile.cohesiv-ci
@@ -1,15 +1,11 @@
-FROM golang:alpine
-
-RUN mkdir -p /go/bin
-ENV GOBIN=/go/bin
-ENV PATH=$PATH:/go/bin
+FROM node:alpine
 
 ENV KUBECTL_VERSION 1.7.4
 ENV KOPS_VERSION 1.7.0
 ENV HELM_VERSION 2.6.0
 
 #install kubectl
-RUN apk add --update \
+RUN apk add --no-cache --update \
   curl \
   jq \
   vim \
