+++ cli/version.go
@@ -1,0 +1,4 @@
+package main
+
+// VERSION is a constant used to checksum and cache cli versions in CI environments
+const VERSION string = "1.0.0"
