--- client/Dockerfile.dev
@@ -1,19 +1,0 @@
-FROM mhart/alpine-node:8
-## git is temporary until react-native-web gets its shit together
-RUN apk add --update git && rm -rf /var/cache/apk/*
-# create app directory in container
-RUN mkdir -p /app/server /app/public
-# set /app directory as default working directory
-WORKDIR /app
-# only copy package.json initially so that `RUN yarn` layer is recreated only
-# if there are changes in package.json
-ADD package.json yarn.lock /app/
-# --pure-lockfile: Don’t generate a yarn.lock lockfile
-RUN yarn --pure-lockfile
-# reproduce project structure
-COPY ./server/ /app/server/
-COPY ./public/ /app/public/
-# expose port 4040
-EXPOSE 4040
-# cmd to start service
-CMD yarn start:server
