+++ client/kube/staging/ingress.yaml
@@ -2,7 +2,7 @@ apiVersion: extensions/v1beta1
 kind: Ingress
 metadata:
   namespace: cohesiv
-  name: mypds
+  name: cohesiv
   annotations:
     kubernetes.io/ingress.class: nginx
     ingress.kubernetes.io/ssl-redirect: "true"
@@ -10,13 +10,13 @@ metadata:
 spec:
   tls:
   - hosts:
-    - mypds.staging.cohesiv.io
+    - cohesiv.staging.cohesiv.io
     secretName: cohesiv-tls
   rules:
-  - host: mypds.staging.cohesiv.io
+  - host: cohesiv.staging.cohesiv.io
     http:
       paths:
       - path: /
         backend:
-          serviceName: mypds
+          serviceName: cohesiv
           servicePort: 80
