+++ client/kube/staging/service.yaml
@@ -2,13 +2,13 @@ apiVersion: v1
 kind: Service
 metadata:
   namespace: cohesiv
-  name: mypds
+  name: cohesiv
   labels:
-    application: mypds
+    application: cohesiv
 spec:
   type: NodePort
   selector:
-    app: mypds
+    app: cohesiv
   ports:
   - protocol: TCP
     port: 80
