+++ client/server/src/tests/auth.test.js
@@ -2,8 +2,8 @@ import chai, { expect } from 'chai'
 import request from 'supertest-as-promised'
 import httpStatus from 'http-status'
 import jwt from 'jsonwebtoken'
-import app from '../dist/index'
-import config from '../dist/config/index'
+import app from '../index'
+import config from '../config/index'
 
 describe('## Auth APIs', () => {
   const validUserCredentials = {
