+++ client/server/tsconfig.json
@@ -5,7 +5,6 @@
     "noImplicitAny": false,
     "moduleResolution": "node",
     "sourceMap": true,
-    "outDir": "dist",
     "allowJs": true,
     "baseUrl": ".",
     "paths": {
