--- configs.yaml
@@ -1,64 +1,0 @@
----
-clients:
- -
-  name: cohesiv
-  develop:
-    deployment:
-    ingress:
-    service:
-  staging:
-    deployment:
-    ingress:
-    service:
-  production:
-    deployment:
-    ingress:
-    service:
- -
-  name: mypds
-  develop:
-    deployment:
-      spec:
-        template:
-          spec:
-            containers:
-            -
-              image: 750213922035.dkr.ecr.us-east-1.amazonaws.com/singularity/mypds:latest
-    ingress:
-    service:
-  staging:
-    deployment:
-      spec:
-        template:
-          spec:
-            containers:
-            -
-              image: 750213922035.dkr.ecr.us-east-1.amazonaws.com/singularity/mypds:latest
-    ingress:
-    service:
-  production:
-    deployment:
-      spec:
-        revisionHistoryLimit: 5
-        replicas: 3
-        template:
-          spec:
-            containers:
-            -
-              image: 750213922035.dkr.ecr.us-east-1.amazonaws.com/singularity/mypds:master
-    ingress:
-    service:
- -
-  name: asu
-  develop:
-    deployment:
-    ingress:
-    service:
-  staging:
-    deployment:
-    ingress:
-    service:
-  production:
-    deployment:
-    ingress:
-    service:
