--- defaults/deployment.yaml
@@ -1,18 +1,0 @@
-apiVersion: extensions/v1beta1
-kind: Deployment
-metadata:
-  namespace: cohesiv
-  name: ${CLIENT}-deployment
-spec:
-  revisionHistoryLimit: 3
-  replicas: 1
-  strategy:
-    rollingUpdate:
-      maxUnavailable: 40%
-  template:
-    metadata:
-      labels:
-        app: ${CLIENT}
-    spec:
-      containers:
-        # ../defaults/container.yaml
