--- defaults/ingress.yaml
@@ -1,22 +1,0 @@
-apiVersion: extensions/v1beta1
-kind: Ingress
-metadata:
-  name: ${CLIENT}
-  namespace: cohesiv
-  annotations:
-    kubernetes.io/ingress.class: nginx
-    ingress.kubernetes.io/ssl-redirect: true
-    kubernetes.io/tls-acme: true
-spec:
-  tls:
-  - hosts:
-    - ${CLIENT}.${HOST_TARGET}cohesiv.io
-    secretName: cohesiv-tls
-  rules:
-  - host: ${CLIENT}.${HOST_TARGET}cohesiv.io
-    http:
-      paths:
-      - path: /
-        backend:
-          serviceName: ${CLIENT}
-          servicePort: 80
