--- defaults/service.yaml
@@ -1,15 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: ${CLIENT}
-  namespace: cohesiv
-  labels:
-    application: ${CLIENT}
-spec:
-  type: NodePort
-  selector:
-    app: ${CLIENT}
-  ports:
-  - protocol: TCP
-    port: 80
-    targetPort: 4040
