+++ installCli.sh
@@ -1,10 +1,5 @@
-IFS=' ' read -r -a array <<< $(go list -f '{{ join .Imports " "}}' github.com/SingularityInteractive/cohesiv)
-for element in "${array[@]}"
-do
-    echo "$element"
-    go get "$element"
-done
-[ -z "$GOBIN" ] && export GOBINZ=$GOPATH/bin
+go build -o $GOPATH/bin/cohesiv github.com/SingularityInteractive/cohesiv/cli
+[ -z "$GOBIN" ] && export GOBIN=$GOPATH/bin
 go install cohesiv.go
 echo "cohesiv Command Line Interface has been installed"
 cohesiv
