+++ templates/app.json
@@ -2,23 +2,23 @@
   "expo": {
     "sdkVersion": "20.0.0",
     "name": "Cohesiv",
-    "slug": "${CLIENT}",
-    "description": "Cohesiv Client",
+    "slug": "{{.Name}}",
+    "description": "{{.Name}} Client",
     "icon": "./assets/icons/app-icon.png",
     "notification": {
-      "icon": "./variants/${CLIENT}/assets/icons/notification-icon.png",
+      "icon": "./variants/{{.Name}}/assets/icons/notification-icon.png",
       "color": "#000000"
     },
     "loading": {
-      "icon": "./variants/${CLIENT}/assets/icons/loading-icon.png",
+      "icon": "./variants/{{.Name}}/assets/icons/loading-icon.png",
       "hideExponentText": true
     },
     "android": {
-      "package": "com.github.SingularityInteractive.${CLIENT}"
+      "package": "com.github.SingularityInteractive.{{.Name}}"
     },
     "ios": {
       "supportsTablet": true,
-      "bundleIdentifier": "com.github.SingularityInteractive.${CLIENT}",
+      "bundleIdentifier": "com.github.SingularityInteractive.{{.Name}}",
       "publishBundlePath": "ios/cohesiv/Supporting/shell-app.bundle",
       "publishManifestPath": "ios/cohesiv/Supporting/shell-app-manifest.json"
     },
