+++ templates/docker-compose.yml
@@ -82,13 +82,13 @@ services:
   client:
     restart: always
     build:
-      context: ./client/builds/mypds
+      context: ./client/builds/{{.Name}}
     command: yarn start:server
     ports:
       - "4040:4040"
     environment:
       NODE_ENV: development
-      DB_CONNECTION_STRING: postgres://postgres:password@postgres:5432/mypds?sslmode=disable
+      DB_CONNECTION_STRING: postgres://postgres:password@postgres:5432/{{.Name}}?sslmode=disable
       PORT: '4040'
       JWT_SECRET: 'abcdefg'
       ACCESS_SVC_ADDR: 'access:8002'
