+++ templates/kube.yaml
@@ -1,0 +1,94 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: {{.Name}}
+  namespace: cohesiv
+  labels:
+    application: {{.Name}}
+spec:
+  type: NodePort
+  selector:
+    app: {{.Name}}
+  ports:
+  - protocol: TCP
+    port: 80
+    targetPort: 4040
+---
+apiVersion: extensions/v1beta1
+kind: Ingress
+metadata:
+  name: {{.Name}}
+  namespace: cohesiv
+  annotations:
+    kubernetes.io/ingress.class: nginx
+    ingress.kubernetes.io/ssl-redirect: true
+    kubernetes.io/tls-acme: true
+spec:
+  tls:
+  - hosts:
+    - {{buildSubdomain .Name .Target.Subdomain}}.cohesiv.io
+    secretName: {{.Name}}-tls
+  rules:
+  - host: {{buildSubdomain .Name .Target.Subdomain}}.cohesiv.io
+    http:
+      paths:
+      - path: /
+        backend:
+          serviceName: {{.Name}}
+          servicePort: 80
+---
+apiVersion: extensions/v1beta1
+kind: Deployment
+metadata:
+  namespace: cohesiv
+  name: {{.Name}}-deployment
+spec:
+  revisionHistoryLimit: {{.Target.RevisionHistoryLimit}}
+  replicas: {{.Target.ReplicaCount}}
+  strategy:
+    rollingUpdate:
+      maxUnavailable: 40%
+  template:
+    metadata:
+      labels:
+        app: {{.Name}}
+    spec:
+      # [START containers]
+      containers:
+      - name: {{.Name}}
+        image: "750213922035.dkr.ecr.us-east-1.amazonaws.com/singularity/{{.Name}}:{{.Target.BranchName}}"
+        imagePullPolicy: Always
+        ports:
+        - containerPort: 4040
+        env:
+        - name: PORT
+          value: "4040"
+        - name: NODE_ENV
+          value: production
+        - name: JWT_SECRET
+          value: fgL8y7-ABYLZyudHBxwUXQaNsNdzJtfYjsUw_cBefJQj8s8G9s1ZWJDmNmmgG1qy
+        - name: TAG_SVC_ADDR
+          valueFrom:
+            configMapKeyRef:
+              name: cohesiv
+              key: hosts.tagdirectory
+        - name: ACCESS_SVC_ADDR
+          valueFrom:
+            configMapKeyRef:
+              name: cohesiv
+              key: hosts.access
+        - name: DB_CONNECTION_STRING
+          valueFrom:
+            secretKeyRef:
+              name: {{.Name}}-secrets
+              key: {{.Target.BranchName}}-db-connection-string
+        resources:
+          requests:
+            cpu: 100m
+            memory: 64Mi
+          limits:
+            memory: 256Mi
+        livenessProbe:
+          initialDelaySeconds: 10
+          tcpSocket:
+            port: 4040
